+++ packages/hasura_cache_interceptor/lib/hasura_cache_interceptor.dart
@@ -1,0 +1,5 @@
+library hasura_cache_interceptor;
+
+export 'src/cache_interceptor.dart';
+export 'src/services/memory_storage_service.dart';
+export 'src/services/storage_service_interface.dart';