+++ packages/hasura_connect/example/lib/app/app_module.dart
@@ -6,7 +6,8 @@ import 'modules/home/home_module.dart';
 class AppModule extends Module {
   @override
   final List<Bind> binds = [
-    Bind.lazySingleton((i) => HasuraConnect('https://bwolfdev.herokuapp.com/v1/graphql')),
+    Bind.lazySingleton(
+        (i) => HasuraConnect('https://bwolfdev.herokuapp.com/v1/graphql')),
   ];
 
   @override