+++ packages/hasura_connect/example/lib/app/modules/home/home_module.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_modular/flutter_modular.dart';
 
-import 'home_Page.dart';
+import 'home_page.dart';
 import 'home_store.dart';
 
 class HomeModule extends Module {