+++ packages/hasura_connect/lib/src/core/interceptors/interceptor.dart
@@ -6,9 +6,9 @@ import 'package:hasura_connect/src/presenter/hasura_connect_base.dart';
 import '../../domain/models/request.dart';
 
 abstract class Interceptor {
-  Future<dynamic>? onRequest(Request request);
-  Future<dynamic>? onResponse(Response data);
-  Future<dynamic>? onError(HasuraError request);
+  Future<dynamic>? onRequest(Request request, HasuraConnect connect);
+  Future<dynamic>? onResponse(Response data, HasuraConnect connect);
+  Future<dynamic>? onError(HasuraError request, HasuraConnect connect);
   Future<void>? onSubscription(Request request, Snapshot snapshot);
   Future<void>? onConnected(HasuraConnect connect);
   Future<void>? onTryAgain(HasuraConnect connect);
@@ -17,27 +17,23 @@ abstract class Interceptor {
 
 abstract class InterceptorBase extends Interceptor {
   @override
-  Future<void>? onConnected(HasuraConnect connect) async {}
+  Future<void> onConnected(HasuraConnect connect) async {}
 
   @override
-  Future<void>? onDisconnected() async {}
+  Future<void> onDisconnected() async {}
 
   @override
-  Future? onError(HasuraError error) async => error;
+  Future onError(HasuraError error, HasuraConnect connect) async => error;
 
   @override
-  Future? onRequest(Request request) async => request;
+  Future onRequest(Request request, HasuraConnect connect) async => request;
 
   @override
-  Future? onResponse(Response data) async => data;
+  Future onResponse(Response data, HasuraConnect connect) async => data;
 
   @override
-  Future<void>? onSubscription(Request request, Snapshot snapshot) {
-    return null;
-  }
+  Future<void> onSubscription(Request request, Snapshot snapshot) async {}
 
   @override
-  Future<void>? onTryAgain(HasuraConnect connect) {
-    return null;
-  }
+  Future<void> onTryAgain(HasuraConnect connect) async {}
 }