+++ packages/hasura_connect/lib/src/domain/models/query.dart
@@ -3,9 +3,10 @@ import 'dart:convert';
 class Query {
   final String document;
   final Map<String, dynamic>? variables;
+  final Map<String, String>? headers;
   final String? key;
 
-  const Query({required this.document, this.variables, this.key});
+  const Query({required this.document, this.variables, this.key, this.headers});
 
   bool get isValid {
     return document.startsWith('query') ||
@@ -14,7 +15,7 @@ class Query {
   }
 
   Map toJson() {
-    return {'query': document, 'variables': variables};
+    return {'query': document, 'variables': variables, 'headers': headers};
   }
 
   @override
@@ -23,11 +24,13 @@ class Query {
   Query copyWith({
     String? document,
     Map<String, dynamic>? variables,
+    Map<String, String>? headers,
     String? key,
   }) {
     return Query(
       document: document ?? this.document,
       variables: variables ?? this.variables,
+      headers: headers ?? this.headers,
       key: key ?? this.key,
     );
   }