+++ packages/hasura_connect/lib/src/presenter/hasura_connect_base.dart
@@ -44,11 +44,11 @@ class HasuraConnect {
   Connector? _connector;
 
   late StreamSubscription _subscription;
-  final int? reconnectionAttemp;
+  final int? reconnectionAttempt;
   final Map<String, String>? headers;
 
   HasuraConnect(this.url,
-      {this.reconnectionAttemp,
+      {this.reconnectionAttempt,
       List<Interceptor>? interceptors,
       this.headers,
       http.Client Function()? httpClientFactory}) {
@@ -97,10 +97,13 @@ class HasuraConnect {
 
   ///Execute a Query from a Document
   Future query(String document,
-      {String? key, Map<String, dynamic>? variables}) async {
+      {String? key,
+      Map<String, dynamic>? variables,
+      Map<String, String>? headers}) async {
     key = key ?? _keyGenerator.generateBase(document);
     return executeQuery(Query(
       key: key,
+      headers: headers,
       document: document.trimLeft(),
       variables: variables,
     ));
@@ -109,14 +112,19 @@ class HasuraConnect {
   ///Execute a Query from a Query
   Future executeQuery(Query query) async {
     final usecase = sl.get<QueryToServer>();
+    var _headers = Map<String, String>.from(headers ?? {});
+    if (query.headers != null) {
+      _headers.addAll(query.headers!);
+    }
+
     var request = Request(
-      headers: headers,
+      headers: _headers,
       type: RequestType.query,
       url: url,
       query: query,
     );
     final interceptedValue = await _interceptorExecutor(
-      ClientResolver.request(request),
+      ClientResolver.request(request, this),
     );
 
     if (interceptedValue is Response) {
@@ -133,7 +141,7 @@ class HasuraConnect {
 
   Future<Response> _interceptError(HasuraError error) async {
     final interceptedValue = await _interceptorExecutor(
-      ClientResolver.error(error),
+      ClientResolver.error(error, this),
     );
 
     if (interceptedValue is Response) {
@@ -145,7 +153,7 @@ class HasuraConnect {
 
   Future<Response> _interceptResponse(Response response) async {
     final interceptedValue = await _interceptorExecutor(
-      ClientResolver.response(response),
+      ClientResolver.response(response, this),
     );
 
     if (interceptedValue is Response) {
@@ -159,11 +167,13 @@ class HasuraConnect {
   Future mutation(String document,
       {Map<String, dynamic>? variables,
       bool tryAgain = true,
-      String? key}) async {
+      String? key,
+      Map<String, String>? headers}) async {
     key = key ?? _keyGenerator.randomString(15);
 
     return executeMutation(Query(
       key: key,
+      headers: headers,
       document: document.trimLeft(),
       variables: variables,
     ));
@@ -173,15 +183,20 @@ class HasuraConnect {
   Future executeMutation(Query query) async {
     final usecase = sl.get<MutationToServer>();
 
+    var _headers = Map<String, String>.from(headers ?? {});
+    if (query.headers != null) {
+      _headers.addAll(query.headers!);
+    }
+
     var request = Request(
-      headers: headers,
+      headers: _headers,
       type: RequestType.mutation,
       url: url,
       query: query,
     );
 
     final interceptedValue = await _interceptorExecutor(
-      ClientResolver.request(request),
+      ClientResolver.request(request, this),
     );
 
     if (interceptedValue is Response) {
@@ -198,12 +213,15 @@ class HasuraConnect {
 
   ///Execute a Subscription from a Document
   Future<Snapshot> subscription(String document,
-      {String? key, Map<String, dynamic>? variables}) async {
+      {String? key,
+      Map<String, dynamic>? variables,
+      Map<String, String>? headers}) async {
     document = document.trim();
     key = key ?? _keyGenerator.generateBase(document);
 
     return executeSubscription(Query(
       key: key,
+      headers: headers,
       document: document,
       variables: variables,
     ));
@@ -280,8 +298,8 @@ class HasuraConnect {
     final connector = _connector!;
     _disconnectionFlag = false;
 
-    if (reconnectionAttemp != null && reconnectionAttemp! > 0) {
-      if (_numbersOfConnectionAttempts >= reconnectionAttemp!) {
+    if (reconnectionAttempt != null && reconnectionAttempt! > 0) {
+      if (_numbersOfConnectionAttempts >= reconnectionAttempt!) {
         print('maximum connection attempt numbers reached');
         _isConnected = false;
         // ignore: unawaited_futures
@@ -300,7 +318,7 @@ class HasuraConnect {
     );
 
     final interceptedValue = await _interceptorExecutor(
-      ClientResolver.request(request),
+      ClientResolver.request(request, this),
     );
 
     try {