+++ packages/hasura_connect/test/src/core/interceptors/interceptor_executor_test.dart
@@ -33,7 +33,7 @@ void main() async {
 
   test('should return same object if interceptor list be empty or null', () {
     var exec = InterceptorExecutor(null);
-    final resolver = ClientResolver.request(request);
+    final resolver = ClientResolver.request(request, HasuraConnectMock());
     expect(exec(resolver), completion(request));
     exec = InterceptorExecutor([]);
     expect(exec(resolver), completion(request));
@@ -43,7 +43,7 @@ void main() async {
     test('should exec interceptor request', () async {
       final exec = InterceptorExecutor(
           [InterceptorMock(onRequestF: (r) async => r.copyWith(url: 'test'))]);
-      final resolver = ClientResolver.request(request);
+      final resolver = ClientResolver.request(request, HasuraConnectMock());
       final result = await exec(resolver);
       expect(result, isA<Request>());
       expect(result.url, 'test');
@@ -52,14 +52,14 @@ void main() async {
     test('should exec interceptor returning other type', () async {
       final exec = InterceptorExecutor(
           [InterceptorMock(onRequestF: (r) async => response)]);
-      final resolver = ClientResolver.request(request);
+      final resolver = ClientResolver.request(request, HasuraConnectMock());
       final result = await exec(resolver);
       expect(result, isA<Response>());
     });
     test('should exec interceptor throw InterceptorError', () async {
       final exec = InterceptorExecutor(
           [InterceptorMock(onRequestF: (r) async => throw Exception('error'))]);
-      final resolver = ClientResolver.request(request);
+      final resolver = ClientResolver.request(request, HasuraConnectMock());
       expect(exec(resolver), throwsA(isA<InterceptorError>()));
     });
   });
@@ -68,7 +68,7 @@ void main() async {
       when(() => response.statusCode).thenReturn(1);
       final exec = InterceptorExecutor(
           [InterceptorMock(onResponseF: (r) async => response)]);
-      final resolver = ClientResolver.response(response);
+      final resolver = ClientResolver.response(response, HasuraConnectMock());
       final result = await exec(resolver);
       expect(result, isA<Response>());
       expect(result.statusCode, 1);
@@ -87,7 +87,7 @@ void main() async {
           ),
         )
       ]);
-      final resolver = ClientResolver.response(response);
+      final resolver = ClientResolver.response(response, HasuraConnectMock());
       final result = await exec(resolver);
       expect(result, isA<HasuraRequestError>());
     });
@@ -96,7 +96,7 @@ void main() async {
         () async {
       final exec = InterceptorExecutor(
           [InterceptorMock(onResponseF: (r) async => request)]);
-      final resolver = ClientResolver.response(response);
+      final resolver = ClientResolver.response(response, HasuraConnectMock());
       expect(exec(resolver), throwsA(isA<InterceptorError>()));
     });
 
@@ -104,7 +104,7 @@ void main() async {
       final exec = InterceptorExecutor([
         InterceptorMock(onResponseF: (r) async => throw Exception('error'))
       ]);
-      final resolver = ClientResolver.response(response);
+      final resolver = ClientResolver.response(response, HasuraConnectMock());
       expect(exec(resolver), throwsA(isA<InterceptorError>()));
     });
   });
@@ -113,7 +113,7 @@ void main() async {
     test('should exec interceptor ', () async {
       final exec =
           InterceptorExecutor([InterceptorMock(onErrorF: (r) async => error)]);
-      final resolver = ClientResolver.error(error);
+      final resolver = ClientResolver.error(error, HasuraConnectMock());
       final result = await exec(resolver);
       expect(result, isA<HasuraError>());
     });
@@ -121,7 +121,7 @@ void main() async {
     test('should exec interceptor returning other type', () async {
       final exec = InterceptorExecutor(
           [InterceptorMock(onErrorF: (r) async => response)]);
-      final resolver = ClientResolver.error(error);
+      final resolver = ClientResolver.error(error, HasuraConnectMock());
       final result = await exec(resolver);
       expect(result, isA<Response>());
     });
@@ -130,14 +130,14 @@ void main() async {
         () async {
       final exec = InterceptorExecutor(
           [InterceptorMock(onErrorF: (r) async => request)]);
-      final resolver = ClientResolver.error(error);
+      final resolver = ClientResolver.error(error, HasuraConnectMock());
       expect(exec(resolver), throwsA(isA<InterceptorError>()));
     });
 
     test('should exec interceptor throw InterceptorError', () async {
       final exec = InterceptorExecutor(
           [InterceptorMock(onErrorF: (r) async => throw Exception('error'))]);
-      final resolver = ClientResolver.error(error);
+      final resolver = ClientResolver.error(error, HasuraConnectMock());
       expect(exec(resolver), throwsA(isA<InterceptorError>()));
     });
   });
@@ -190,13 +190,16 @@ class InterceptorMock extends Interceptor {
       this.onRequestF,
       this.onResponseF});
   @override
-  Future<dynamic>? onError(HasuraError error) => onErrorF?.call(error);
+  Future<dynamic>? onError(HasuraError error, HasuraConnect connect) =>
+      onErrorF?.call(error);
 
   @override
-  Future? onRequest(Request request) => onRequestF?.call(request);
+  Future? onRequest(Request request, HasuraConnect connect) =>
+      onRequestF?.call(request);
 
   @override
-  Future? onResponse(Response response) => onResponseF?.call(response);
+  Future? onResponse(Response response, HasuraConnect connect) =>
+      onResponseF?.call(response);
 
   @override
   Future<void>? onConnected(HasuraConnect connect) =>