+++ packages/hasura_connect/test/src/infra/repositories/request_repository_test.dart
@@ -20,7 +20,7 @@ void main() {
   final tRequest =
       Request(url: '', query: Query(document: 'query', key: 'dadas'));
 
-  registerFallbackValue<Request>(tRequest);
+  registerFallbackValue(tRequest);
 
   setUpAll(() {
     response = ResponseMock();